- Fixed errors and warnings when compiling with GCC. (Unfortunately, the VC++ debug builds   become ungodly slow when using mods with complex DECORATE. The GCC debug builds run just   fine, however. Hopefully this is something that can be fixed later with an assembly-optimized   version of the main VM loop, because I don't relish the thought of being stuck with GDB   for debugging.) - Fixed: The ACS_Named* action specials were erroneously defined as taking strings instead of   names. - Fixed: Copy-paste error caused FxMultiNameState::Emit to generate code that called   DecoNameToClass instead of DecoFindMultiNameState. - Updated FxActionSpecialCall::Emit for named script specials. - Fixed inverted asserts for FxMinusSign::Emit and FxUnaryNotBitwise::Emit.
SVN r3893 (scripting)

+++ src/g_shared/a_pickups.cpp
@@ -1273,7 +1273,7 @@ bool AInventory::DoRespawn ()
 void AInventory::GiveQuest (AActor *toucher)
 {
 	int quest = GetClass()->GiveQuest;
-	if (quest > 0 && quest <= countof(QuestItemClasses))
+	if (quest > 0 && quest <= (int)countof(QuestItemClasses))
 	{
 		toucher->GiveInventoryType (QuestItemClasses[quest-1]);
 	}
